refactor: separate MessageIcon logic from useMessageHelpers for better react rule-following
